pyproject.toml
[project]
name = "construction-submittal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.76.0",
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "llama-index>=0.14.13",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-anthropic>=0.10.6",
    "llama-index-vector-stores-chroma>=0.5.5",
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]
